Resolve conflicts and complete Page model refactor to Pydantic discriminated union
- Merged latest changes from main (LLM logic moved to summarize_other.py).
- Refactored `Page` in `core.py` to a Pydantic discriminated union of `ArxivPage` and `OtherPage`.
- `ArxivPage` now includes `paper_id`, `abstract`, and strict `YYYY` date format.
- `OtherPage` maintains existing `YYYY/MM/DD` or `????/??/??` date formats.
- Added `parse_page` helper with backward compatibility for missing `kind` field.
- Updated `summarize_other.py` to populate specialized arXiv fields.
- Updated `send_to_personal.py`, `send_to_queue_v2.py`, and `add_summary_to_eval_queue.py` to use `parse_page`.
- Updated `.gitignore` to exclude `__pycache__` and `*.pyc`.
