+++ src/endpoint/readit/core.py
@@ -1,20 +1,9 @@
-from langchain_google_genai import ChatGoogleGenerativeAI
-from langchain_core.prompts import ChatPromptTemplate
 from pydantic import BaseModel
 from pydantic import Field
 from pydantic import HttpUrl
 from pydantic import TypeAdapter
 from pydantic import BeforeValidator
 from typing import Any, Literal, Annotated, Union
-
-
-class Summary(BaseModel):
-    title: str = Field(
-        description="The title of concise main title of the article or page"
-    )
-    date: str = Field(
-        description="The issue or publication date as YYYY/MM/DD format (????/??/?? if unknown)"
-    )
 
 
 class FetchResult(BaseModel):
@@ -61,29 +50,3 @@ PageAdapter = TypeAdapter(Annotated[Page, BeforeValidator(_default_kind_to_other
 
 def parse_page(data: dict[str, Any]) -> Page:
     return PageAdapter.validate_python(data)
-
-
-_PROMPT = ChatPromptTemplate.from_template("""
-    Analyze the following content from a webpage and extract two pieces of information:
-    1. The concise main title of the article or page.
-    2. The issue or publication date as YYYY/MM/DD format (if available).
-       - If not available, state "????/??/??".
-
-    Format your answer as a JSON object with keys "date" and "title".
-
-    Content: {content}
-    """)
-_STRUCTURED_LLM = ChatGoogleGenerativeAI(
-    model="gemini-2.5-flash"
-).with_structured_output(Summary)
-_CHAIN = _PROMPT | _STRUCTURED_LLM
-
-
-def page_of_(fetch_result: FetchResult) -> OtherPage:
-    summary = _CHAIN.invoke({"content": fetch_result.html})
-
-    return OtherPage(
-        url=fetch_result.url,
-        title=summary.title,
-        date=summary.date,
-    )
